cli: move exec commands into each topic's own folder since they diverge
