+++ packages/cli/src/commands/xcode/create.ts
@@ -61,27 +61,32 @@ export default class XcodeCreate extends BaseCommand {
 
       const start = Date.now();
       const instance = await this.client.xcodeInstances.create(params);
+      const consoleUrl = this.consoleStreamUrl(instance.metadata.id);
       saveLastInstanceId(instance.metadata.id);
-      this.log(`Created a new Xcode instance in ${((Date.now() - start) / 1000).toFixed(1)}s`);
-      this.log(`Instance ID: ${instance.metadata.id}`);
-      this.log(`Region: ${instance.spec.region}`);
-      this.log(`State: ${instance.status.state}`);
+      this.info(`Created a new Xcode instance in ${((Date.now() - start) / 1000).toFixed(1)}s`);
+      this.info('Xcode Instance:');
+      this.info(`  ID: ${instance.metadata.id}`);
+      this.info(`  Console URL: ${consoleUrl}`);
+      this.info(`  Region: ${instance.spec.region}`);
+      this.info(`  State: ${instance.status.state}`);
 
       if (flags.json) {
         this.outputJson(instance);
+      } else if (this.isQuietEnabled()) {
+        this.output(instance.metadata.id);
       }
 
       if (flags.rm) {
         const cleanup = async () => {
           try {
             await this.client.xcodeInstances.delete(instance.metadata.id);
-            this.log(`${instance.metadata.id} is deleted`);
+            this.info(`${instance.metadata.id} is deleted`);
           } catch (e) {
-            this.log(`Failed to delete instance: ${e}`);
+            this.info(`Failed to delete instance: ${e}`);
           }
         };
 
-        this.log('Instance running. Press Ctrl+C to stop and delete.');
+        this.info('Instance running. Press Ctrl+C to stop and delete.');
         await new Promise<void>((resolve) => {
           const keepAlive = setInterval(() => {}, 1 << 30);
           const shutdown = () => {
